Refactor dist tests
* Split up AbstractJanusGraphAssemblyIT into a generic class for all release tests and suite for default backend tests
* Activate single vertex tests
* Centralized naming in pom.xml for zip names
