+++ janusgraph-dist/pom.xml
@@ -24,10 +24,14 @@
         <assembly.es.dir>${project.build.directory}/elasticsearch-${elasticsearch.version}</assembly.es.dir>
         <assembly.cassandra.dir>${project.build.directory}/apache-cassandra-${cassandra.version}</assembly.cassandra.dir>
 
+        <packname.standard>janusgraph-${project.version}</packname.standard>
+        <packname.full>janusgraph-full-${project.version}</packname.full>
+
         <!-- Variables for integration testing on release-candidate archive files -->
         <test.expect.dir>${project.build.directory}/test-classes</test.expect.dir>
-        <test.zipfile.path>${project.build.directory}/janusgraph-${project.version}.zip</test.zipfile.path>
-        <test.zipfile-full.path>${project.build.directory}/janusgraph-full-${project.version}.zip</test.zipfile-full.path>
+        <test.zipfile.path>${project.build.directory}/${packname.standard}.zip</test.zipfile.path>
+        <test.zipfile-full.path>${project.build.directory}/${packname.full}.zip</test.zipfile-full.path>
+        <build.full>package</build.full>
 
         <!-- RPM/DEB packaging -->
         <pkg.release>1</pkg.release>
@@ -674,7 +678,7 @@
                         <executions>
                             <execution>
                                 <id>assemble-full-distribution-archive</id>
-                                <phase>package</phase>
+                                <phase>${build.full}</phase>
                                 <goals>
                                     <goal>single</goal>
                                 </goals>
@@ -689,7 +693,7 @@
                                         ${finalName}-${<id /> from assembly descriptor}.${archive
                                         format extension}
                                     -->
-                                    <finalName>janusgraph-full-${project.version}</finalName>
+                                    <finalName>${packname.full}</finalName>
                                     <appendAssemblyId>false</appendAssemblyId>
                                     <tarLongFileMode>gnu</tarLongFileMode>
                                     <descriptors>
@@ -709,14 +713,7 @@
                                 <configuration>
                                     <!--
                                         We set attached=true so that the gpg plugin will
-                                        sign all of our assembly artifacts. We handle
-                                        deployment of the artifacts manually because
-                                        maven-deploy-plugin can't upload files to a flat
-                                        S3 directory; it insists on creating a
-                                        long/packagename/and/artifactid/and/version/
-                                        path at the destination before uploading, and we
-                                        don't want that for our files S3. codehaus's
-                                        wagon:upload plugin comes to our rescue.
+                                        sign all of our assembly artifacts.
                                     -->
                                     <attach>true</attach>
                                     <!--
@@ -724,7 +721,7 @@
                                         ${finalName}-${<id /> from assembly descriptor}.${archive
                                         format extension}
                                     -->
-                                    <finalName>janusgraph-${project.version}</finalName>
+                                    <finalName>${packname.standard}</finalName>
                                     <appendAssemblyId>false</appendAssemblyId>
                                     <tarLongFileMode>gnu</tarLongFileMode>
                                     <descriptors>
