+++ janusgraph-dist/src/test/java/org/janusgraph/pkgtest/AbstractJanusGraphAssemblyIT.java
@@ -14,25 +14,7 @@
 
 package org.janusgraph.pkgtest;
 
-import static org.apache.tinkerpop.gremlin.driver.ser.AbstractMessageSerializer.TOKEN_IO_REGISTRIES;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertTimeout;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.PrintStream;
-import java.io.Writer;
-import java.net.Socket;
-import java.time.Duration;
-import java.util.*;
-
-import org.apache.commons.io.FileUtils;
+import com.google.common.collect.ImmutableMap;
 import org.apache.tinkerpop.gremlin.driver.Cluster;
 import org.apache.tinkerpop.gremlin.driver.MessageSerializer;
 import org.apache.tinkerpop.gremlin.driver.remote.DriverRemoteConnection;
@@ -44,102 +26,88 @@ import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSo
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.apache.tinkerpop.gremlin.structure.io.graphson.GraphSONMapper;
 import org.apache.tinkerpop.gremlin.structure.io.gryo.GryoMapper;
-import org.apache.velocity.Template;
-import org.apache.velocity.VelocityContext;
-import org.apache.velocity.app.Velocity;
-
-import com.google.common.base.Joiner;
-import com.google.common.collect.ImmutableMap;
 import org.janusgraph.graphdb.tinkerpop.JanusGraphIoRegistry;
+import org.junit.jupiter.api.Test;
 
-public abstract class AbstractJanusGraphAssemblyIT {
+import java.net.Socket;
+import java.time.Duration;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
-    protected static final String BUILD_DIR;
-    protected static final String EXPECT_DIR;
-    protected static final String ZIPFILE_PATH;
-    protected static final String ZIPFILE_EXTRACTED;
-    protected static final String ZIPFILE_FULL_PATH;
-    protected static final String ZIPFILE_FULL_EXTRACTED;
+import static org.apache.tinkerpop.gremlin.driver.ser.AbstractMessageSerializer.TOKEN_IO_REGISTRIES;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertTimeout;
 
-    static {
-        Properties props;
+public abstract class AbstractJanusGraphAssemblyIT extends JanusGraphAssemblyBaseIT {
 
-        try {
-            props = new Properties();
-            java.io.FileReader fr = new FileReader(Joiner.on(File.separator).join(new String[] { "target", "test-classes", "target.properties" }));
-            props.load(fr);
-            fr.close();
-        } catch (IOException e) {
-            throw new AssertionError(e);
-        }
+    protected abstract String getConfigPath();
 
-        BUILD_DIR    = props.getProperty("build.dir");
-        EXPECT_DIR   = props.getProperty("expect.dir");
-        ZIPFILE_PATH = props.getProperty("zipfile.path");
-        ZIPFILE_EXTRACTED = ZIPFILE_PATH.substring(0, ZIPFILE_PATH.length() - 4);
-        ZIPFILE_FULL_PATH = props.getProperty("zipfile-full.path");
-        ZIPFILE_FULL_EXTRACTED = ZIPFILE_FULL_PATH.substring(0, ZIPFILE_FULL_PATH.length() - 4);
+    protected abstract String getServerConfigPath();
 
-        Properties p = new Properties();
-        p.put("file.resource.loader.path", EXPECT_DIR);
-        Velocity.init(p);
+    protected abstract String getGraphName();
+
+    @Test
+    public void testSingleVertexInteractionAgainstGremlinSh() throws Exception {
+        unzipAndRunExpect("single-vertex.expect.vm", getConfigPath(), getGraphName(), false, false);
     }
 
-    protected void testSimpleGremlinSession(String graphConfig, String graphToString, boolean full) throws Exception {
-        unzipAndRunExpect("single-vertex.expect.vm", graphConfig, graphToString, full, false);
+    @Test
+    public void testSingleVertexInteractionAgainstGremlinShFull() throws Exception {
+        unzipAndRunExpect("single-vertex.expect.vm", getConfigPath(), getGraphName(), true, false);
     }
 
-    private static boolean serverListening(String host, int port)
-    {
-        Socket s = null;
-        try
-        {
-            s = new Socket(host, port);
-            return true;
-        }
-        catch (Exception e)
-        {
-            return false;
-        }
-        finally
-        {
-            if(s != null)
-                try {s.close();}
-                catch(Exception e){}
-        }
+    @Test
+    public void testGettingStartedAgainstGremlinSh() throws Exception {
+        unzipAndRunExpect("getting-started.expect.vm", getConfigPath(), getGraphName(), false, false);
     }
 
-    protected void testJanusGraphServer(String janusgraphServerConfig, boolean full) throws Exception {
-        final boolean debug = false;
-        ImmutableMap<String, String> contextVars = ImmutableMap.of("janusgraphServerConfig", janusgraphServerConfig);
-        unzipAndRunExpect("janusgraph-server-sh.before.expect.vm", contextVars, full, debug);
-        assertTimeout(Duration.ofSeconds(30), () -> {
-            while (!serverListening("localhost",8182)) {
-                Thread.sleep(1000);
-            }
-        });
-
-        runTraversalAgainstServer(createGraphSONMessageSerializer());
-        runTraversalAgainstServer(createGraphBinaryMessageSerializerV1());
-        runTraversalAgainstServer(createGryoMessageSerializer());
-
-        parseTemplateAndRunExpect("janusgraph-server-sh.after.expect.vm", contextVars, full, debug);
+    @Test
+    public void testGettingStartedAgainstGremlinShFull() throws Exception {
+        unzipAndRunExpect("getting-started.expect.vm", getConfigPath(), getGraphName(), true, false);
     }
 
-    private MessageSerializer createGryoMessageSerializer() {
+    @Test
+    public void testJanusGraphServerGremlin() throws Exception {
+        testJanusGraphServer(false);
+    }
+
+    @Test
+    public void testJanusGraphServerGremlinFull() throws Exception {
+        testJanusGraphServer(true);
+    }
+
+    protected MessageSerializer createGryoMessageSerializer() {
         return new GryoMessageSerializerV3d0(GryoMapper.build().addRegistry(JanusGraphIoRegistry.instance()));
     }
 
-    private MessageSerializer createGraphSONMessageSerializer() {
+    protected MessageSerializer createGraphSONMessageSerializer() {
         return new GraphSONMessageSerializerV3d0(GraphSONMapper.build().addRegistry(JanusGraphIoRegistry.instance()));
     }
 
-    private MessageSerializer createGraphBinaryMessageSerializerV1() {
+    protected MessageSerializer createGraphBinaryMessageSerializerV1() {
         final GraphBinaryMessageSerializerV1 serializer = new GraphBinaryMessageSerializerV1();
         final Map<String, Object> config = new HashMap<>();
         config.put(TOKEN_IO_REGISTRIES, Collections.singletonList(JanusGraphIoRegistry.class.getName()));
         serializer.configure(config, Collections.emptyMap());
         return serializer;
+    }
+
+    protected static boolean serverListening(String host, int port) {
+        Socket s = null;
+        try {
+            s = new Socket(host, port);
+            return true;
+        } catch (Exception e) {
+            return false;
+        } finally {
+            if (s != null)
+                try {
+                    s.close();
+                } catch (Exception e) {
+                }
+        }
     }
 
     private void testServerUsingTraversalSource(GraphTraversalSource g) {
@@ -148,7 +116,7 @@ public abstract class AbstractJanusGraphAssemblyIT {
         assertNotEquals(0, vertices.size());
     }
 
-    private void runTraversalAgainstServer(MessageSerializer serializer) {
+    protected void runTraversalAgainstServer(MessageSerializer serializer) {
         Cluster cluster = Cluster.build("localhost")
             .port(8182)
             .serializer(serializer)
@@ -160,131 +128,20 @@ public abstract class AbstractJanusGraphAssemblyIT {
         testServerUsingTraversalSource(g);
     }
 
-    protected void testGettingStartedGremlinSession(String graphConfig, String graphToString, boolean full) throws Exception {
-        unzipAndRunExpect("getting-started.expect.vm", graphConfig, graphToString, full, false);
-    }
-
-    protected void unzipAndRunExpect(String expectTemplateName, Map<String, String> contextVars, boolean full, boolean debug) throws Exception {
-        if (full) {
-            FileUtils.deleteQuietly(new File(ZIPFILE_FULL_EXTRACTED));
-            unzip(BUILD_DIR, ZIPFILE_FULL_PATH);
-        } else {
-            FileUtils.deleteQuietly(new File(ZIPFILE_EXTRACTED));
-            unzip(BUILD_DIR, ZIPFILE_PATH);
-        }
-
-        parseTemplateAndRunExpect(expectTemplateName, contextVars, full, debug);
-    }
-
-    protected void parseTemplateAndRunExpect(String expectTemplateName, Map<String, String> contextVars) throws IOException, InterruptedException {
-        parseTemplateAndRunExpect(expectTemplateName, contextVars, true, false);
-    }
-
-    protected void parseTemplateAndRunExpect(String expectTemplateName, Map<String, String> contextVars, boolean full, boolean debug) throws IOException, InterruptedException {
-        VelocityContext context = new VelocityContext();
-        for (Map.Entry<String, String> ent : contextVars.entrySet()) {
-            context.put(ent.getKey(), ent.getValue());
-        }
-
-        Template template = Velocity.getTemplate(expectTemplateName);
-        String inputPath = EXPECT_DIR + File.separator + expectTemplateName;
-        String outputPath = inputPath.substring(0, inputPath.length() - 3);
-
-        Writer output = new FileWriter(outputPath);
-        template.merge(context, output);
-        output.close();
-
-        if (full) {
-            expect(ZIPFILE_FULL_EXTRACTED, outputPath, debug);
-        } else {
-            expect(ZIPFILE_EXTRACTED, outputPath, debug);
-        }
-    }
-
-    protected void unzipAndRunExpect(String expectTemplateName, boolean full) throws Exception {
-        unzipAndRunExpect(expectTemplateName, Collections.emptyMap(), full, false);
-    }
-
-    protected void unzipAndRunExpect(String expectTemplateName, String graphConfig, String graphToString, boolean full, boolean debug) throws Exception {
-        unzipAndRunExpect(expectTemplateName, ImmutableMap.of("graphConfig", graphConfig, "graphToString", graphToString), full, debug);
-    }
-
-    private static void expect(String dir, String expectScript, boolean debug) throws IOException, InterruptedException {
-        if (debug) {
-            command(new File(dir), "expect", "-d", expectScript);
-        } else {
-            command(new File(dir), "expect", expectScript);
-        }
-    }
-
-    protected static void unzip(String dir, String zipFile) throws IOException, InterruptedException {
-        command(new File(dir), "unzip", "-q", zipFile);
-    }
-
-    protected static void command(File dir, String... command) throws IOException, InterruptedException {
-        ProcessBuilder pb = new ProcessBuilder(command);
-        pb.directory(dir);
-//        pb.redirectInput(Redirect.PIPE);
-        /*
-         * Using Redirect.INHERIT with expect breaks maven-failsafe-plugin when
-         * failsafe is configured to fork. The parent and child normally
-         * communicate over stdout/stderr in fork mode. But after executing
-         * expect, this failsafe communication starts appearing on the terminal.
-         * The parent never sees it and assumes no tests ran. expect is probably
-         * doing something nasty to its file descriptors and neglecting to clean
-         * up after itself. Invoking unzip does not break failsafe+forks in this
-         * way. So expect must be doing something unusual with its file
-         * descriptors.
-         *
-         * Redirect.INHERIT works fine if failsafe is configured to never fork.
-         */
-//        pb.redirectOutput(Redirect.INHERIT);
-//        pb.redirectError(Redirect.INHERIT);
-//        pb.redirectOutput(Redirect.PIPE);
-//        pb.redirectError(Redirect.PIPE);
-        final Process p = pb.start();
-        // Sense of "input" and "output" are reversed between ProcessBuilder and Process
-        p.getOutputStream().close(); // Child process sees EOF on stdin (if it reads stdin at all)
-        Thread outPrinter = new Thread(new SubprocessPipePrinter(p.getInputStream(), System.out));
-        Thread errPrinter = new Thread(new SubprocessPipePrinter(p.getErrorStream(), System.out));
-        outPrinter.start();
-        errPrinter.start();
-        int stat = p.waitFor();
-        outPrinter.join();
-        errPrinter.join();
-        assertEquals(0, stat);
-    }
-
-    private static class SubprocessPipePrinter implements Runnable {
-
-        private final BufferedReader source;
-        private final PrintStream sink;
-
-        private SubprocessPipePrinter(InputStream source, PrintStream sink) {
-            this.source = new BufferedReader(new InputStreamReader(source));
-            this.sink = sink;
-        }
-
-        @Override
-        public void run() {
-            try {
-                runUnsafe();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            } catch (InterruptedException e) {
-                // Exit silently
+    protected void testJanusGraphServer(boolean full) throws Exception {
+        final boolean debug = false;
+        ImmutableMap<String, String> contextVars = ImmutableMap.of("janusgraphServerConfig", getServerConfigPath());
+        unzipAndRunExpect("janusgraph-server-sh.before.expect.vm", contextVars, full, debug);
+        assertTimeout(Duration.ofSeconds(30), () -> {
+            while (!serverListening("localhost", 8182)) {
+                Thread.sleep(1000);
             }
-        }
+        });
 
-        private void runUnsafe() throws IOException, InterruptedException {
+        runTraversalAgainstServer(createGraphSONMessageSerializer());
+        runTraversalAgainstServer(createGraphBinaryMessageSerializerV1());
+        runTraversalAgainstServer(createGryoMessageSerializer());
 
-            String line;
-
-            while (null != (line = source.readLine())) {
-                synchronized (sink) {
-                    sink.println(line);
-                }
-            }
-        }
+        parseTemplateAndRunExpect("janusgraph-server-sh.after.expect.vm", contextVars, full, debug);
     }
 }
