+++ janusgraph-dist/src/test/java/org/janusgraph/pkgtest/BerkeleyAssemblyIT.java
@@ -14,27 +14,32 @@
 
 package org.janusgraph.pkgtest;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 public class BerkeleyAssemblyIT extends AbstractJanusGraphAssemblyIT {
+    @Override
+    protected String getConfigPath() {
+        return "conf/janusgraph-berkeleyje.properties";
+    }
 
-    @Test
-    public void testSimpleGremlinSession() throws Exception {
-        testSimpleGremlinSession("conf/janusgraph-berkeleyje.properties", "berkeleyje", false);
+    @Override
+    protected String getServerConfigPath() {
+        return "conf/gremlin-server/gremlin-server-berkeleyje.yaml";
+    }
+
+    @Override
+    protected String getGraphName() {
+        return "berkeleyje";
     }
 
     @Test
-    public void testSimpleGremlinSessionFull() throws Exception {
-        testSimpleGremlinSession("conf/janusgraph-berkeleyje.properties", "berkeleyje", true);
-    }
+    @Disabled
+    @Override
+    public void testGettingStartedAgainstGremlinSh() {}
 
     @Test
-    public void testJanusGraphServer() throws Exception {
-        testJanusGraphServer("conf/gremlin-server/gremlin-server-berkeleyje.yaml", false);
-    }
-
-    @Test
-    public void testJanusGraphServerFull() throws Exception {
-        testJanusGraphServer("conf/gremlin-server/gremlin-server-berkeleyje.yaml", true);
-    }
+    @Disabled
+    @Override
+    public void testGettingStartedAgainstGremlinShFull() {}
 }
