+++ janusgraph-dist/src/test/java/org/janusgraph/pkgtest/BerkeleyESAssemblyIT.java
@@ -15,7 +15,6 @@
 package org.janusgraph.pkgtest;
 
 import org.janusgraph.diskstorage.es.JanusGraphElasticsearchContainer;
-import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
@@ -23,24 +22,20 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 public class BerkeleyESAssemblyIT extends AbstractJanusGraphAssemblyIT {
 
     @Container
-    private static JanusGraphElasticsearchContainer esr = new JanusGraphElasticsearchContainer(true);
+    private static JanusGraphElasticsearchContainer es = new JanusGraphElasticsearchContainer(true);
 
-    @Test
-    public void testSimpleGremlinSession() throws Exception {
-        testGettingStartedGremlinSession("conf/janusgraph-berkeleyje-es.properties", "berkeleyje", false);
-    }
-    @Test
-    public void testSimpleGremlinSessionFull() throws Exception {
-        testGettingStartedGremlinSession("conf/janusgraph-berkeleyje-es.properties", "berkeleyje", true);
+    @Override
+    protected String getConfigPath() {
+        return "conf/janusgraph-berkeleyje-es.properties";
     }
 
-    @Test
-    public void testJanusGraphServer() throws Exception {
-        testJanusGraphServer("conf/gremlin-server/gremlin-server-berkeleyje-es.yaml", false);
+    @Override
+    protected String getServerConfigPath() {
+        return "conf/gremlin-server/gremlin-server-berkeleyje-es.yaml";
     }
 
-    @Test
-    public void testJanusGraphServerFull() throws Exception {
-        testJanusGraphServer("conf/gremlin-server/gremlin-server-berkeleyje-es.yaml", true);
+    @Override
+    protected String getGraphName() {
+        return "berkeleyje";
     }
 }
